Added given step to create folder with files
